spread tests and utilities over multiple files
